pyproject.toml
[tool.poetry]
name = "connebts"
version = "0.1.0"
description = ""
authors = ["Grivtsov_e"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
fastapi = "^0.71.0"
uvicorn = "^0.16.0"
pydantic = { extras = ["dotenv"], version = "^1.8.2" }
Hypercorn = "^0.11.2"
uvloop = "^0.15.2"
asyncpg = "^0.23.0"
aiocron = "^1.8"
starlette-exporter = "^0.12.0"
sentry-sdk = "^1.5.8"
sse-starlette = "^0.10.3"
python-jose = "^3.3.0"
sqlalchemy-utils = "^0.38.3"
passlib = "^1.7.4"
bcrypt = "^4.0.0"

[tool.poetry.dev-dependencies]
isort = "^5.8.0"
black = "22.3.0"
flake8 = "^3.9.2"
mypy = "^0.902"
pytest = "^6.2.5"
pytest-asyncio = "^0.17.2"
pytest-mock = "3.6.1"
asgi-lifespan = "^1.0.1"
alembic = "^1.6.5"
psycopg2-binary = "^2.9.1"
sqlalchemy = "^1.4.20"
pylint = "^2.10.2"
requests = "2.26.0"
pytest-httpx = "^0.18.0"
pytest-cov = "^3.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
py38 = true
line_length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
     venv
    | \.git
    | alembic
    )/
)
'''

[[tool.poetry.source]]
name = "pypi-proxy"
url = "https://nxsdct.tass.ru/repository/tass-pypi-proxy/simple"
default = true

[[tool.poetry.source]]
name = "tasslib"
url = "https://tass_pypi:PAxLCUVAYWhJzZKGfycM@gitlab.corp.tass.ru/api/v4/projects/146/packages/pypi/simple"
secondary = true
